+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Refactor API errors handling methods to `api/errors.ts` file
 
 ### v0.73.6
 * `@types/blueimp-md5` is a dev dependency, not an actual dependency