+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # NEXT RELEASE
 * Refactor API errors handling methods to `api/errors.ts` file
-* Add logStringifiedProps() to Logger
+* Add `logStringifiedProps()` to `Logger` class
 
 ### v0.73.6
 * `@types/blueimp-md5` is a dev dependency, not an actual dependency